+++ src/main/scala/info/gratour/jt808common/codec/decoder/JT808MsgBodyDecoderRegistry.scala
@@ -14,3 +14,8 @@ object DefaultMsgBodyDecoderRegistry extends JT808MsgBodyDecoderRegistry {
   private final val Map: Map[Int, JT808MsgBodyDecoder[_]] = JT808MsgBodyDecoders.ALL.map(d => d.msgId -> d).toMap
   override def get(msgId: Int): JT808MsgBodyDecoder[_] = Map.get(msgId).orNull
 }
+
+object SimplifiedMsgBodyDecoderRegistry extends JT808MsgBodyDecoderRegistry {
+  private final val Map: Map[Int, JT808MsgBodyDecoder[_]] = JT808MsgBodyDecoders.ALL_SIMPLIFIED.map(d => d.msgId -> d).toMap
+  override def get(msgId: Int): JT808MsgBodyDecoder[_] = Map.get(msgId).orNull
+}